+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Add admin component SignIn
 - Add admin component SignUp
 - Add demo admin component AdminHeader
+- Add demo admin component SideMenu
 ### Changed
 - Updated page store with action returning a promise
 ### Deprecated